[IMP] models: allow prefetching of fields in draft mode but not in onchange mode
Change the condition to disable the prefetching in onchange mode only.  The
prefetching of fields was disabled in draft mode in order to handle the case of
onchanges on records: computed fields must to be recomputed and not fetched
from database.

This saves a few queries when computing a field.  For instance, viewing a
question on the forum takes 66 queries instead of 85.